requirements.txt
fastmcp>=0.1.0
pillow>=10.0.0
